Implented create, delete, fetch all, fetch by id endpoints for product entity

+++ LabWebAPI.Services/Services/ProductService.cs
@@ -6,6 +6,7 @@ using LabWebAPI.Contracts.DTO.Product;
 using LabWebAPI.Contracts.Exceptions;
 using LabWebAPI.Contracts.Roles;
 using LabWebAPI.Contracts.Services;
+using LabWebAPI.Services.Validators;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using System;
@@ -22,40 +23,52 @@ namespace LabWebAPI.Services.Services
         private readonly IMapper _mapper;
         private readonly IRepository<Product> _productRepository;
         private readonly RoleManager<IdentityRole> _roleManager;
+        private readonly UserManager<User> _userManager;
 
         public ProductService(IMapper mapper,
         IRepository<Product> productRepository,
+        UserManager<User> userManager,
         RoleManager<IdentityRole> roleManager)
         {
             _mapper = mapper;
             _productRepository = productRepository;
             _roleManager = roleManager;
+            _userManager = userManager;
         }
-    //    public async Task<bool> CanUserDeleteProductAsync(string userId, int productId)
-    //    {
-     //       var product = await _productRepository.GetByKeyAsync(productId);
-    //        return product != null && (product.UserWhoCreated.Id == userId || IsAdmin(userId));
-//}
 
-        public async Task DeleteProductAsync(string id)
+
+        public async Task DeleteProductAsync(string id, string userId)
         {
             var product = await _productRepository.GetByKeyAsync(id) ?? throw new ProductNotFoundException("Product not found!");
-            await _productRepository.DeleteAsync(product);
-            await _productRepository.SaveChangesAsync();
+            var _isAdmin = await Validator.IsAdmin(_userManager, userId);
+            if (_isAdmin || product.UserWhoCreatedId == userId)
+            {
+                await _productRepository.DeleteAsync(product);
+                await _productRepository.SaveChangesAsync();
+            }
+            else
+            {
+                throw new UnauthorizedToProductChangeException("Unauthorized To Product Delete");
+            }
+        
         }
         public async Task<ProductDTO> GetProductByIdAsync(string id)
         {
-           
-            var product = await _productRepository.GetByKeyAsync(id) ?? throw new ProductNotFoundException("Product not found!");
+
+            var product = await _productRepository.Query()
+       .Include(p => p.UserWhoCreated)
+       .FirstOrDefaultAsync(p => p.Id == id)
+       ?? throw new ProductNotFoundException("Product not found!");
+
             var model = new ProductDTO();
-            _mapper.Map(product, model);
-          
+           _mapper.Map(product, model);
+
             return model;
         }
 
         public async Task<IEnumerable<ProductDTO>> GetProductsAsync()
         {
-            var products = await _productRepository.GetAllAsync();
+            var products = await _productRepository.Query().Include(p => p.UserWhoCreated).ToListAsync();
             var productsInfo = products.Select(product =>
             {
 
@@ -66,8 +79,14 @@ namespace LabWebAPI.Services.Services
                     Description = product.Description,
 
                     Price = product.Price,
-                    PublicationDate = product.PublicationDate
-                    
+                    PublicationDate = product.PublicationDate,
+                      UserWhoCreated = new SimpleUserInfoDTO
+                      {
+                          Id = product.UserWhoCreated.Id,
+                          Name = product.UserWhoCreated.Name,
+                          Email = product.UserWhoCreated.Email, 
+                          Surname = product.UserWhoCreated.Surname
+                      }
                 };
             })
             .ToList();
@@ -80,6 +99,27 @@ namespace LabWebAPI.Services.Services
 
             return model;
         }
+
+        public async Task<bool> AddProduct(Product model)
+        {
+
+            try
+            {
+
+                // Додати продукт до репозиторію
+                await _productRepository.AddAsync(model);
+
+                // Зберегти зміни в базі даних
+                await _productRepository.SaveChangesAsync();
+
+                return true;
+            }
+            catch (Exception ex)
+            {
+                return false;
+            }
+        }
+
 
     }
